martian: move Proxy.connect() to proxy_connect.go
